+++ src/utilities/filterContants.js
@@ -1,0 +1,59 @@
+export const genderList = [
+  {
+    value: '0',
+    label: 'Male'
+  },
+  {
+    value: '1',
+    label: 'Female'
+  },
+  {
+    value: 'unisex',
+    label: 'Unisex'
+  }
+];
+
+export const prices = [
+  {
+    value: { min: 0, max: 25000 },
+    label: '< ₹25,000'
+  },
+  {
+    value: { min: 25000, max: 50000 },
+    label: '₹25,000 - ₹50,000'
+  },
+  {
+    value: { min: 50000, max: 100000 },
+    label: '₹50,000 - ₹1,00,000'
+  },
+  {
+    value: { min: 100000 },
+    label: '₹1,00,000+'
+  }
+];
+
+export const purityList = [
+  {
+    label: '14K',
+    value: '14'
+  },
+  {
+    label: '18K',
+    value: '18'
+  },
+  {
+    label: '22K',
+    value: '22'
+  },
+  {
+    label: '92.5K',
+    value: '92.5'
+  }
+];
+
+ export const sortbyList = [
+    'popularity',
+    'new arrival',
+    'price low to high',
+    'price high to low'
+  ];
